[clang][contracts] Fix warnings in the contracts implementation
Warnings pass (final-passes increment 2) over the contracts-p3850 changes,
mirroring the GCC pass.  Addresses every warning attributable to our changes
(both the efcs and p3850 epochs), including warnings induced on pre-existing
upstream lines by our shared-entity changes.

- -Woverflow (x109): getFutureCompatDiagKind() narrowed the now-64-bit keyword
  flags to unsigned (KEYCONTRACTSP3400 truncated to 0).  Widen the StringSwitch
  and its local to uint64_t; clears all keyword and BuiltinTraits.inc overflow
  warnings.
- -Wswitch (x8): handle the P4298 NoexceptObserve/NoexceptEnforce values of
  ContractEvaluationSemantic in SemanticAsString and the TextNodeDumper switch,
  and add their spellings to the ContractSemantic attribute's enum argument
  (regenerates the covered dumper switches).
- -Wreturn-type (x3): add llvm_unreachable after the covered switches in
  GetContractViolationDescriptorTypeString and the two CGContracts IIFEs.
- -Wc++20-extensions (x~25): replace designated initializers with per-field
  assignment (ContractScopeRecord, CurrentContractInfo).
- -Wunused-but-set-variable: drop the unused KWLoc in the contract_control
  primary-expression parse.
- -Wempty-body: remove the dead no-op 'if (isInExportDeclContext()) ;'.
- -Wmissing-prototypes (x20): declare the P4298 _noexcept ABI entry points in
  <__contracts/abi.h>.

+++ clang/lib/Basic/IdentifierTable.cpp
@@ -837,7 +837,7 @@ IdentifierTable::getFutureCompatDiagKind(const IdentifierInfo &II,
                                          const LangOptions &LangOpts) {
   assert(II.isFutureCompatKeyword() && "diagnostic should not be needed");
 
-  unsigned Flags = llvm::StringSwitch<unsigned>(II.getName())
+  uint64_t Flags = llvm::StringSwitch<uint64_t>(II.getName())
 #define KEYWORD(NAME, FLAGS) .Case(#NAME, FLAGS)
 #include "clang/Basic/TokenKinds.def"
 #undef KEYWORD
